+++ components/Accordion/constants/src/index.ts
@@ -13,41 +13,3 @@ export const CssClasses = {
   contentWrapper: 'border border-gray-100 rounded-b-md border-t-0',
   content: 'px-[24px] py-[16px]',
 }
-
-export interface AccordionProps {
-  /**
-   * Main indigo title.
-   * [NOTE] Its color and font can be customized using `titleClassName`.
-   */
-  title: string
-  /**
-   * Second line in the heading.
-   */
-  description?: string
-  /**
-   * Icon to be displayed on the left of the the heading.
-   */
-  icon?: any
-  /**
-   * Should we add a vertical separator between the icon and the text.
-   */
-  separator?: boolean
-  /**
-   * Change the font and color of the heading title
-   */
-  titleClassName?: string
-  /**
-   * Change the font and color of the heading description
-   */
-  descriptionClassName?: string
-  /**
-   * Additional classes to add to the header of the accordion
-   * > [NOTE] useful to change the background color of the header
-   */
-  headingClassName?: string
-  /**
-   * When using content that needs ti be edge to edge,
-   * removes the content wrapper from the content.
-   */
-  fullWidthContent?: boolean
-}
